aider: Added meta tag to specify UTF-8 character encoding.
# Aider chat conversation:

USER: The emojis are coming out like "ðŸ˜€" please fix

ASSISTANT: The issue you're experiencing is likely due to the file encoding. Emojis are Unicode characters, and if the file is not being served with the correct encoding, they can appear as gibberish.

To fix this, we need to specify the character encoding in the HTML file. This is done by adding a meta tag in the head of the HTML document.

Here is the change that needs to be made:

```html
index.html
<<<<<<< HEAD
<head>
=======
<head>
    <meta charset="UTF-8">
>>>>>>> updated
```

This change tells the browser to interpret the document as UTF-8, which supports all Unicode characters including emojis.