aider: Added sound and animation for matching and non-matching cards.
# Aider chat conversation:

USER: After flipping two cards, the game pauses: if correct, play applause.wav, if incorrect, shake the cards and flip them back over.

ASSISTANT: To implement this feature, we need to modify the JavaScript code in the `index.html` file. We will add code to play the `applause.wav` sound when a match is found and to shake the cards and flip them back over when no match is found.

Here are the changes that need to be made:

```html
index.html
<<<<<<< HEAD
            // Check for a match
            var flippedCards = document.querySelectorAll('button:not([data-matched])');
            if (flippedCards.length === 2) {
                if (flippedCards[0].dataset.emoji === flippedCards[1].dataset.emoji) {
                    // Match found
                    flippedCards[0].dataset.matched = 'true';
                    flippedCards[1].dataset.matched = 'true';
                } else {
                    // No match, flip the cards back over
                    setTimeout(function() {
                        flippedCards[0].textContent = '';
                        flippedCards[1].textContent = '';
                    }, 1000);
                }
            }
=======
            // Check for a match
            var flippedCards = document.querySelectorAll('button:not([data-matched])');
            if (flippedCards.length === 2) {
                if (flippedCards[0].dataset.emoji === flippedCards[1].dataset.emoji) {
                    // Match found
                    flippedCards[0].dataset.matched = 'true';
                    flippedCards[1].dataset.matched = 'true';
                    // Play applause sound
                    var audio = new Audio('applause.wav');
                    audio.play();
                } else {
                    // No match, shake the cards and flip them back over
                    flippedCards[0].classList.add('shake');
                    flippedCards[1].classList.add('shake');
                    setTimeout(function() {
                        flippedCards[0].textContent = '';
                        flippedCards[1].textContent = '';
                        flippedCards[0].classList.remove('shake');
                        flippedCards[1].classList.remove('shake');
                    }, 1000);
                }
            }
>>>>>>> updated
```

This change adds the `Audio` object to play the `applause.wav` sound when a match is found. It also adds the `shake` class to the cards when no match is found, which will cause them to shake. After 1 second, the cards are flipped back over and the `shake` class is removed.

Please note that you need to have the `applause.wav` sound file in the same directory as your `index.html` file for the sound to play. Also, you need to define the `shake` class in your CSS for the shake animation to work. If you need help with these, please let me know.