aider: Added a 3-second delay before hiding the matched cards.
# Aider chat conversation:

USER: When a match happens, it should pause 3s before hiding the cards.

ASSISTANT: To add a delay before hiding the matched cards, we can use the `setTimeout` function in JavaScript. This function allows us to delay the execution of a piece of code.

We need to modify the `index.html` file.

Here are the steps:

1. In the `flipCard` function, when a match is found, use `setTimeout` to delay the addition of the `.matched` class by 3 seconds.

Here are the changes:

```html
index.html
<<<<<<< HEAD
                        flippedCards[0].classList.add('matched');
                        flippedCards[1].classList.add('matched');
=======
                        setTimeout(function() {
                            flippedCards[0].classList.add('matched');
                            flippedCards[1].classList.add('matched');
                        }, 3000);
>>>>>>> updated
```

These changes will add a 3-second delay before hiding the matched cards.