aider: Set font-size to 100% for card emojis.
# Aider chat conversation:

USER: Make the emojis as large as they can be in their containing grid cell.

ASSISTANT: To make the emojis as large as they can be in their containing grid cell, we can use CSS. We can set the `font-size` to a percentage value, which will make the size of the emojis relative to the size of their parent element.

We need to modify the `index.html` file.

Here are the steps:

1. In the CSS rule for `.card.flipped::before, .card.matched::before`, set `font-size` to `100%`.

Here are the changes:

```html
index.html
<<<<<<< HEAD
        .card.flipped::before, .card.matched::before {
            content: attr(data-emoji);
        }
=======
        .card.flipped::before, .card.matched::before {
            content: attr(data-emoji);
            font-size: 100%;
        }
>>>>>>> updated
```

These changes will make the emojis as large as they can be in their containing grid cell.