aider: Moved clearing of `flippedCards` array into the `setTimeout` function.
# Aider chat conversation:

USER: Now for a match, nothing happens at all, even after 3s

ASSISTANT: The issue might be due to the fact that the `flippedCards` array is being cleared immediately after the `setTimeout` function is set up, but before it has a chance to execute. This means that by the time the `setTimeout` function runs, the `flippedCards` array is already empty, so there are no cards to add the `matched` class to.

To fix this, we can move the line that clears the `flippedCards` array into the `setTimeout` function, so that it only runs after the `matched` class has been added to the cards.

We need to modify the `index.html` file.

Here are the steps:

1. In the `flipCard` function, move the line that clears the `flippedCards` array into the `setTimeout` function.

Here are the changes:

```html
index.html
<<<<<<< HEAD
                        setTimeout(function() {
                            if (flippedCards.includes(flippedCards[0]) && flippedCards.includes(flippedCards[1])) {
                                flippedCards[0].classList.add('matched');
                                flippedCards[1].classList.add('matched');
                            }
                        }, 3000);
                        flippedCards = [];
=======
                        setTimeout(function() {
                            if (flippedCards.includes(flippedCards[0]) && flippedCards.includes(flippedCards[1])) {
                                flippedCards[0].classList.add('matched');
                                flippedCards[1].classList.add('matched');
                                flippedCards = [];
                            }
                        }, 3000);
>>>>>>> updated
```

These changes will ensure that the `flippedCards` array is not cleared until after the `matched` class has been added to the cards.