aider: Increase the font-size of emojis to 300%.
# Aider chat conversation:

USER: the emojis should be 3x larger

ASSISTANT: To make the emojis 3 times larger, we need to increase the `font-size` in the CSS rule for `.card.flipped::before, .card.matched::before`. Currently, it's set to `100%`, so we'll change it to `300%`.

Here are the changes:

```html
index.html
<<<<<<< HEAD
        .card.flipped::before, .card.matched::before {
            content: attr(data-emoji);
            font-size: 100%;
        }
=======
        .card.flipped::before, .card.matched::before {
            content: attr(data-emoji);
            font-size: 300%;
        }
>>>>>>> updated
```

This change will make the emojis appear 3 times larger when a card is flipped or matched.