aider: Make the cards 3 times larger.
# Aider chat conversation:

USER: make the cards 3x larger

ASSISTANT: To make the cards 3 times larger, we need to modify the CSS rules for the `.card` class in the `index.html` file. Specifically, we need to change the `width` and `height` properties from `100px` to `300px`, and the `font-size` from `2em` to `6em`.

Here is the *edit block* for this change:

```html
index.html
<<<<<<< HEAD
        .card {
            width: 100px;
            height: 100px;
            background: #ccc;
            display: flex;
            justify-content: center;
            align-items: center;
            font-size: 2em;
        }
=======
        .card {
            width: 300px;
            height: 300px;
            background: #ccc;
            display: flex;
            justify-content: center;
            align-items: center;
            font-size: 6em;
        }
>>>>>>> updated
```

Please apply this change to your `index.html` file.