aider: Hide matched cards from the board 3 seconds after a match has been made.
# Aider chat conversation:

USER: Instead of waiting for audio, wait 3s after match before hiding the matched cards from the board

ASSISTANT: To implement this change, we need to modify the `flipCard` function in the `index.html` file. Specifically, we need to replace the `audio.onended` event handler with a `setTimeout` function that waits for 3 seconds before hiding the matched cards.

Here is the *edit block* for this change:

```html
index.html
<<<<<<< HEAD
                    audio.play();
                    audio.onended = function() {
                        hideMatchedCards();
                        flippedCards = [];
                        isProcessing = false;
                        checkWin();
                    }
=======
                    audio.play();
                    setTimeout(function() {
                        hideMatchedCards();
                        flippedCards = [];
                        isProcessing = false;
                        checkWin();
                    }, 3000);
>>>>>>> updated
```

This change will ensure that the matched cards are hidden 3 seconds after they are matched, regardless of the duration of the audio.