aider: Change the visibility of matched cards to 'hidden' instead of changing the `data-emoji` attribute.
# Aider chat conversation:

USER: They are still disappearing as soon as a match happens. Fix it.

ASSISTANT: I apologize for the confusion. It seems there might be an issue with the `hideMatchedCards` function. Currently, it's setting the `data-emoji` attribute of the matched cards to an empty string, which might be causing them to disappear immediately.

Instead, we should only change the visibility of the matched cards after the 3 seconds delay. We can achieve this by modifying the `hideMatchedCards` function to change the visibility of the matched cards to 'hidden' instead of changing the `data-emoji` attribute.

Here is the *edit block* for this change:

```html
index.html
<<<<<<< HEAD
        // Function to hide matched cards
        function hideMatchedCards() {
            const matchedCards = document.querySelectorAll('.card.matched');
            matchedCards.forEach(card => {
                card.dataset.emoji = '';
            });
        }
=======
        // Function to hide matched cards
        function hideMatchedCards() {
            const matchedCards = document.querySelectorAll('.card.matched');
            matchedCards.forEach(card => {
                card.style.visibility = 'hidden';
            });
        }
>>>>>>> updated
```

This change will ensure that the matched cards are hidden 3 seconds after they are matched, regardless of the duration of the audio.