aider: Changed the timing of the `hideMatchedCards` function call to occur after the audio has finished playing.
# Aider chat conversation:

USER: They are still disappearing immediately

ASSISTANT: I apologize for the confusion earlier. It seems like the issue might be with the timing of the `hideMatchedCards` function call. Currently, it's being called immediately after the audio starts playing, which might be causing the cards to disappear immediately.

We should call the `hideMatchedCards` function after the audio has finished playing. We can achieve this by adding an 'ended' event listener to the audio, which will call the `hideMatchedCards` function when the audio has finished playing.

Here is the *edit block* for this change:

```html
index.html
<<<<<<< HEAD
                    audio.play();
                    setTimeout(function() {
                        hideMatchedCards();
                        flippedCards = [];
                        isProcessing = false;
                        checkWin();
                    }, 3000);
=======
                    audio.play();
                    audio.onended = function() {
                        setTimeout(function() {
                            hideMatchedCards();
                            flippedCards = [];
                            isProcessing = false;
                            checkWin();
                        }, 3000);
                    };
>>>>>>> updated
```

This change will ensure that the matched cards are hidden 3 seconds after the audio has finished playing, not immediately when a match is made.