aider: Delay adding the 'matched' class to the cards until after the 3 seconds delay.
# Aider chat conversation:

USER: They are still disappearing immediately. Please fix it. Make sure that when a match happens, no classes are changed on the matched cards until the 3s delay

ASSISTANT: I see. The issue seems to be that the 'matched' class is being added to the cards immediately after a match is found, which is causing them to disappear immediately. We should delay adding the 'matched' class until after the 3 seconds delay.

Here is the *edit block* for this change:

```html
index.html
<<<<<<< HEAD
                    if (flippedCards[0].dataset.emoji === flippedCards[1].dataset.emoji) {
                        flippedCards[0].classList.add('matched');
                        flippedCards[1].classList.add('matched');
                        audio.play();
                        audio.onended = function() {
                            setTimeout(function() {
                                hideMatchedCards();
                                flippedCards = [];
                                isProcessing = false;
                                checkWin();
                            }, 3000);
                        };
=======
                    if (flippedCards[0].dataset.emoji === flippedCards[1].dataset.emoji) {
                        audio.play();
                        audio.onended = function() {
                            setTimeout(function() {
                                flippedCards[0].classList.add('matched');
                                flippedCards[1].classList.add('matched');
                                hideMatchedCards();
                                flippedCards = [];
                                isProcessing = false;
                                checkWin();
                            }, 3000);
                        };
>>>>>>> updated
```

This change will ensure that the 'matched' class is added to the cards 3 seconds after a match is found, not immediately when a match is made.