aider: Add viewport meta tag to optimize page layout for mobile and disable zooming.
# Aider chat conversation:

USER: I get weird encoding bugs when this is hosted on S3

ASSISTANT: This could be due to the character encoding settings on your S3 bucket. S3 uses the `Content-Type` metadata field to determine how to handle the file, and this includes the character encoding. If this is not set correctly, it could cause encoding issues when the file is served.

To fix this, you need to set the `Content-Type` metadata field to `text/html; charset=UTF-8` for your HTML files in your S3 bucket. This will tell S3 to serve the files as HTML with UTF-8 encoding.

Unfortunately, I can't make these changes for you as they need to be done in the S3 management console or using the AWS CLI, and I don't have access to your AWS account. However, I can guide you through the process.

Here are the steps to update the `Content-Type` metadata field for a file in the S3 management console:

1. Open the S3 management console in your web browser.
2. Navigate to your bucket and then to the file you want to update.
3. Click on the file to open its details.
4. Click on the `Metadata` tab.
5. Click `Edit`.
6. In the `Key` field, enter `Content-Type`.
7. In the `Value` field, enter `text/html; charset=UTF-8`.
8. Click `Save`.

You'll need to repeat these steps for each HTML file in your bucket.

If you have the AWS CLI installed, you can also use the `s3api put-object` command to update the metadata for a file. Here's an example command:

```bash
aws s3api put-object --bucket my-bucket --key index.html --body index.html --content-type "text/html; charset=UTF-8"
```

Replace `my-bucket` with the name of your bucket and `index.html` with the path to your HTML file. This command will upload the file to S3 and set the `Content-Type` metadata field at the same time.

Please let me know if you need further assistance with this.

USER: Make sure the page layout is optimized for mobile and zooming is disabled

ASSISTANT: To optimize the page layout for mobile and disable zooming, we need to add a viewport meta tag in the head section of the HTML file. The viewport meta tag controls the layout on mobile browsers. By setting the width to device-width, the layout will adapt to the screen width of the device. By setting the initial-scale to 1 and user-scalable to no, we disable zooming.

Here is the change we need to make:

```html
index.html
<<<<<<< HEAD
<head>
    <title>Memory Game</title>
=======
<head>
    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
    <title>Memory Game</title>
>>>>>>> updated
```

This change will add the viewport meta tag right after the opening head tag.