aider: Removed visible JavaScript comment and function definition.
# Aider chat conversation:

USER: There is a javascript comment and function definition that appear to be visible; please remove them

ASSISTANT: To remove the visible JavaScript comment and function definition, we need to modify the `index.html` file.

Here are the changes:

```html
index.html
<<<<<<< HEAD
        // Function to hide matched cards
        function hideMatchedCards() {
            const matchedCards = document.querySelectorAll('.card.matched');
            matchedCards.forEach(card => {
                card.style.visibility = 'hidden';
            });
        }
=======
>>>>>>> updated
```

This edit block removes the JavaScript comment and function definition that were mistakenly placed outside of the `<script>` tag.