aider: Disable pinch-to-zoom on Safari on iPad by setting maximum-scale to 1 in the viewport meta tag.
# Aider chat conversation:

USER: Can still pinch to zoom on Safari on iPad, we need to disable that -- it's accidentally triggered during game play

ASSISTANT: To disable pinch-to-zoom on Safari on iPad, we need to add `maximum-scale=1` to the viewport meta tag in the HTML file. This will prevent users from zooming in or out on the page.

Here is the change:

```html
index.html
<<<<<<< HEAD
    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
=======
    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no">
>>>>>>> updated
```

This change will set the maximum zoom level to 1, which effectively disables zooming. The `user-scalable=no` part already attempts to disable zooming, but it's not always respected, especially on iOS devices. Adding `maximum-scale=1` is a more reliable way to disable zooming.